+++ mythtv/libs/libmythui/mythuitext.cpp
@@ -704,7 +704,7 @@ bool MythUIText::GetNarrowWidth(const QStringList & paragraphs,
             too_narrow = std::max<qreal>(too_narrow, width);
 
             // Too narrow?  How many lines didn't fit?
-            qreal lines = roundf((height - m_drawRect.height()) / line_height);
+            qreal lines = round((height - m_drawRect.height()) / line_height);
             lines -= (1.0 - last_line_width / width);
             width += (lines * width) /
                 ((double)m_drawRect.height() / line_height);
